new CLI version, currently under development
Now using the newst crypt_guard crate, the only now missing feature that soon is finished would be signature verification/ opening. though encryption & decryption with all 3 Kyber key sizes and both AES and XChaCh20 are are working as well as signing with falcon or dilithium either creating a detached signature or detacheddata. Signing is also not fully finished, while the code is finished it just wasn't thoroughly tested.